Fixed so status check ignores TYPE field
These binaries have been fixed so that they ignore the TYPE field of LSWS's status file. This fix does not bypass or modify any of LSWS's license checks and does not circumvent any DRM covered by the DMCA and similar laws.

The modified binaries in this commit are ones of OpenLiteSpeed, a program that is released under the GPLv3 license. The source code to make these binaries is not provided by the original publisher of it (usmannasir). These modified binaries were made by directly modifying the originals and not by compiling from source code, thus there is no source code for me to distribute with them.

> "(...) the GNU GPL does not restrict what people do in software, it just stops them from restricting others." -GNU